+++ backend/src/main/java/erp/backend/domain/emp/controller/EmpController.java
@@ -2,7 +2,8 @@ package erp.backend.domain.emp.controller;
 
 import erp.backend.domain.emp.dto.SignInRequest;
 import erp.backend.domain.emp.dto.SignInResponse;
-import erp.backend.domain.emp.service.UserService;
+import erp.backend.domain.emp.dto.SignUpRequest;
+import erp.backend.domain.emp.service.EmpService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -15,14 +16,15 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 @RequestMapping("/api/emp")
 public class EmpController {
-    private final UserService userService;
     private final EmpService empService;
 
+    // 로그인
     @PostMapping("/sign-in")
     public ResponseEntity<SignInResponse> signIn(@RequestBody @Valid SignInRequest request) {
-        System.out.println(request.toString());
-        return ResponseEntity.ok(userService.signIn(request));
+        return ResponseEntity.ok(empService.signIn(request));
+    }
 
+    // 회원가입
     @PostMapping("/sign-up")
     public ResponseEntity<Boolean> signUp(@RequestBody @Valid SignUpRequest request) {
         empService.signUp(request);
