+++ backend/src/main/java/erp/backend/domain/emp/service/EmpService.java
@@ -1,24 +1,31 @@
 package erp.backend.domain.emp.service;
 
+import erp.backend.domain.emp.dto.SignInRequest;
+import erp.backend.domain.emp.dto.SignInResponse;
 import erp.backend.domain.emp.dto.SignUpRequest;
 import erp.backend.domain.emp.entity.Emp;
 import erp.backend.domain.emp.repository.EmpRepository;
+import erp.backend.global.config.security.jwt.JwtProvider;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Arrays;
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 public class EmpService {
     private final EmpRepository empRepository;
     private final PasswordEncoder passwordEncoder;
+    private final JwtProvider jwtProvider;
 
     @Transactional
     public void signUp(SignUpRequest request) {
-        // 신입 사원은 초기 비밀번호가 1541로 설정
-        // String password = passwordEncoder.encode(request.getPassword());
 
+        // 신입 사원은 초기 비밀번호가 1541로 설정
         Emp emp = Emp.builder()
                 .empName(request.getEmpName())
                 .empEmail(request.getEmpEmail())
@@ -35,6 +42,30 @@ public class EmpService {
                 .empStatus("재직중")
                 .build();
         empRepository.save(emp);
+    }
+
+    @Transactional(readOnly = true)
+    public SignInResponse signIn(SignInRequest request) {
+        Emp emp = empRepository.findByEmpEmail(request.getEmpEmail())
+                .orElseThrow(() -> new UsernameNotFoundException("이메일/비밀번호가 맞지않습니다."));
+        String encPassword = emp.getPassword();
+        boolean matches = passwordEncoder.matches(request.getPassword(), encPassword);
+        if (!matches) {
+            throw new UsernameNotFoundException("이메일/비밀번호가 맞지않습니다.");
+        }
+
+        String token = jwtProvider.createToken(emp.getEmpEmail(), emp.getAuthorities());
+
+        List<String> roles = Arrays.stream(emp.getRoles().split(",")).toList();
+
+        return SignInResponse.builder()
+                .token(token)
+                .empId(emp.getEmpId())
+                .empName(emp.getEmpName())
+                .empEmail(emp.getEmpEmail())
+                .roles(roles)
+                .build();
+
     }
 
 }
