+++ backend/src/main/resources/docs/DEPT/DEPT.sql
@@ -1,2 +1,7 @@
-DROP TABLE IF EXISTS `final`.DEPT;
-
+DROP TABLE IF EXISTS `final`.dept;
+CREATE TABLE `final`.dept
+(
+    `DEPT_ID`   BIGINT       NOT NULL COMMENT '부서번호',
+    `DEPT_NAME` VARCHAR(255) NOT NULL COMMENT '부서명',
+    PRIMARY KEY (`DEPT_ID`)
+) ENGINE = InnoDB COMMENT '부서';
