+++ backend/src/main/resources/docs/MOVIE/SERVICEMOVIE.sql
@@ -13,4 +13,4 @@ CREATE TABLE `final`.SERVICEMOVIE
   AUTO_INCREMENT = 1;
 
 ALTER TABLE `final`.SERVICEMOVIE
-    ADD CONSTRAINT `FK_SERVICEMOVIE_EMP` FOREIGN KEY (`SERVICEMOVIE_MOVIE_ID`) REFERENCES `final`.MOVIE (`MOVIE_ID`);
+    ADD CONSTRAINT `FK_SERVICEMOVIE_EMP` FOREIGN KEY (`SERVICEMOVIE_MOVIE_ID`) REFERENCES `final`.movie (`MOVIE_ID`);
